move chat to separate folder
