+++ app_utils/sections/experiment.py
@@ -1,9 +1,7 @@
-import gc
 import glob
 import logging
 import os
 import shutil
-import threading
 import zipfile
 from typing import Callable, List, Optional, Set
 
@@ -14,16 +12,14 @@ import pandas as pd
 import torch
 import transformers
 import yaml
-from accelerate import dispatch_model
-from accelerate.utils import get_balanced_memory, infer_auto_device_map
 from h2o_wave import Q, data, ui
 from jinja2 import Environment, FileSystemLoader
 from sqlitedict import SqliteDict
 
 from app_utils.config import default_cfg
+from app_utils.sections.chat import chat_tab, load_cfg_model_tokenizer
 from app_utils.sections.common import clean_dashboard
 from app_utils.utils import (
-    WaveChatStreamer,
     add_model_type,
     flatten_dict,
     get_cfg_list_items,
@@ -42,8 +38,6 @@ from app_utils.utils import (
     start_experiment,
 )
 from app_utils.wave_utils import busy_dialog, ui_table_from_df, wave_theme
-from llm_studio.src.datasets.text_utils import get_tokenizer
-from llm_studio.src.models.text_causal_language_modeling_model import Model
 from llm_studio.src.tooltips import tooltips
 from llm_studio.src.utils.config_utils import (
     convert_cfg_to_nested_dictionary,
@@ -64,13 +58,10 @@ from llm_studio.src.utils.export_utils import (
     save_prediction_outputs,
 )
 from llm_studio.src.utils.logging_utils import write_flag
-from llm_studio.src.utils.modeling_utils import load_checkpoint, unwrap_model
+from llm_studio.src.utils.modeling_utils import unwrap_model
 from llm_studio.src.utils.utils import add_file_to_zip, kill_child_processes
 
 logger = logging.getLogger(__name__)
-
-USER = True
-BOT = False
 
 
 async def experiment_start(q: Q) -> None:
@@ -1108,210 +1099,6 @@ async def logs_tab(q):
     q.client.delete_cards.add("experiment/display/logs")
 
 
-async def chat_tab(q: Q, load_model=True):
-    running_experiments = get_experiments(q=q)
-    running_experiments = running_experiments[
-        running_experiments.status.isin(["running"])
-    ]
-
-    # gpu id in UI is offset by 1 to be in sync with experiment UI
-    gpu_id = q.client["gpu_used_for_chat"] - 1
-    gpu_blocked = any(
-        [
-            str(gpu_id) in gpu_list
-            for gpu_list in running_experiments["gpu_list"]
-            .apply(lambda x: x.split(","))
-            .to_list()
-        ]
-    )
-    if gpu_blocked:
-        q.page["experiment/display/chat"] = ui.form_card(
-            box="first",
-            items=[
-                ui.text(
-                    f"""Chatbot is not available when GPU{q.client["gpu_used_for_chat"]}
-                     is blocked by another experiment.
-                     You can change "Gpu used for Chat" in the settings tab
-                     to use another GPU for the chatbot. """
-                )
-            ],
-            title="",
-        )
-        q.client.delete_cards.add("experiment/display/chat")
-        return
-
-    if load_model:
-        loading_message = "Loading the model..."
-    else:
-        loading_message = "Chat History cleaned. How can I help you?"
-
-    q.page["experiment/display/chat"] = ui.chatbot_card(
-        box="first",
-        data=data(fields="content from_user", t="list"),
-        name="experiment/display/chat/chatbot",
-    )
-    q.page["experiment/display/chat"].data += [loading_message, BOT]
-    q.client["experiment/display/chat/messages"] = []
-    q.client.delete_cards.add("experiment/display/chat")
-
-    q.page["experiment/display/chat/settings"] = ui.form_card(
-        box="second",
-        items=[
-            ui.expander(
-                name="chat_settings",
-                label="Chat Settings",
-                items=[ui.progress(label="Loading the model...")],
-                expanded=True,
-            )
-        ],
-    )
-    q.client.delete_cards.add("experiment/display/chat/settings")
-
-    await q.page.save()
-    logger.info(torch.cuda.memory_allocated())
-
-    if load_model:
-        cfg, model, tokenizer = load_cfg_model_tokenizer(
-            q.client["experiment/display/experiment_path"], device=f"cuda:{gpu_id}"
-        )
-        q.client["experiment/display/chat/cfg"] = cfg
-        q.client["experiment/display/chat/model"] = model
-        q.client["experiment/display/chat/tokenizer"] = tokenizer
-        initial_message = "Model successfully loaded, how can I help you?"
-
-    else:
-        cfg = q.client["experiment/display/chat/cfg"]
-        assert q.client["experiment/display/chat/model"] is not None
-        assert q.client["experiment/display/chat/tokenizer"] is not None
-        # Do not update loading_message.
-        initial_message = loading_message
-
-    # Hide fields that are should not be visible in the UI
-    cfg.prediction._visibility["metric"] = -1
-    cfg.prediction._visibility["batch_size_inference"] = -1
-    cfg.prediction._visibility["min_length_inference"] = -1
-    cfg.prediction._visibility["stop_tokens"] = -1
-
-    logger.info(torch.cuda.memory_allocated())
-    q.page["experiment/display/chat"].data[0] = [
-        initial_message,
-        BOT,
-    ]
-
-    option_items = get_ui_elements(
-        cfg=q.client["experiment/display/chat/cfg"].prediction,
-        q=q,
-        pre="chat/cfg_predictions",
-    )
-    q.page["experiment/display/chat/settings"] = ui.form_card(
-        box="second",
-        items=[
-            ui.button(
-                name="experiment/display/chat/clear_history",
-                label="Clear History",
-                primary=True,
-            ),
-            ui.expander(
-                name="chat_settings",
-                label="Chat Settings",
-                items=option_items,
-                expanded=True,
-            ),
-        ],
-    )
-
-
-@torch.inference_mode(mode=True)
-async def chat_update(q: Q) -> None:
-    """
-    Update the chatbot with the new message.
-    """
-    cfg_prediction = parse_ui_elements(
-        cfg=q.client["experiment/display/chat/cfg"].prediction,
-        q=q,
-        pre="chat/cfg_predictions/cfg/",
-    )
-    logger.info(f"Using chatbot config: {cfg_prediction}")
-    q.client["experiment/display/chat/cfg"].prediction = cfg_prediction
-
-    prompt = q.client["experiment/display/chat/chatbot"]
-
-    message = [prompt, USER]
-    q.client["experiment/display/chat/messages"].append(message)
-    q.page["experiment/display/chat"].data += message
-    q.page["experiment/display/chat"].data += ["", BOT]
-    await q.page.save()
-
-    cfg = q.client["experiment/display/chat/cfg"]
-    model: Model = q.client["experiment/display/chat/model"]
-    tokenizer = q.client["experiment/display/chat/tokenizer"]
-
-    full_prompt = ""
-    if len(q.client["experiment/display/chat/messages"]):
-        for prev_message in q.client["experiment/display/chat/messages"][
-            -(cfg.prediction.num_history + 1) :
-        ]:
-            if prev_message[1] is USER:
-                prev_message = cfg.dataset.dataset_class.parse_prompt(
-                    cfg, prev_message[0]
-                )
-            else:
-                prev_message = prev_message[0]
-                if cfg.dataset.add_eos_token_to_answer:
-                    prev_message += cfg._tokenizer_eos_token
-
-            full_prompt += prev_message
-    logger.info(f"Full prompt: {full_prompt}")
-    inputs = cfg.dataset.dataset_class.encode(
-        tokenizer, full_prompt, cfg.tokenizer.max_length_prompt, "left"
-    )
-    inputs["prompt_input_ids"] = (
-        inputs.pop("input_ids").unsqueeze(0).to(cfg.environment._device)
-    )
-    inputs["prompt_attention_mask"] = (
-        inputs.pop("attention_mask").unsqueeze(0).to(cfg.environment._device)
-    )
-
-    def text_cleaner(text: str) -> str:
-        return cfg.dataset.dataset_class.clean_output(
-            output={"predicted_text": np.array([text])}, prompts=[full_prompt], cfg=cfg
-        )["predicted_text"][0]
-
-    if cfg.prediction.num_beams == 1:
-        streamer = WaveChatStreamer(tokenizer=tokenizer, q=q, text_cleaner=text_cleaner)
-    else:
-        # ValueError: `streamer` cannot be used with beam search (yet!).
-        # Make sure that `num_beams` is set to 1.
-        streamer = None
-        logger.info("Not streaming output, as it cannot be used with beam search.")
-
-    output = {}
-
-    def generate(inputs, cfg, streamer):
-        with torch.cuda.amp.autocast():
-            output["predicted_answer_ids"] = (
-                model.generate(batch=inputs, cfg=cfg, streamer=streamer).detach().cpu()
-            )
-
-    threading.Thread(
-        target=generate, kwargs=dict(inputs=inputs, cfg=cfg, streamer=streamer)
-    ).start()
-
-    predicted_text = streamer.answer
-    predicted_text = text_cleaner(predicted_text)
-    logger.info(f"Predicted Answer: {predicted_text}")
-
-    message = [predicted_text, BOT]
-    q.client["experiment/display/chat/messages"].append(message)
-    q.page["experiment/display/chat"].data[-1] = message
-
-    del output
-    del inputs
-
-    gc.collect()
-    torch.cuda.empty_cache()
-
-
 def subsample(key1, key2, value, max_plot_points=1000):
     act_plot_points = len(value["steps"])
     if act_plot_points > max_plot_points:
@@ -1902,54 +1689,3 @@ def get_model_card(cfg, model, repo_id) -> huggingface_hub.ModelCard:
         else "",
     )
     return card
-
-
-def load_cfg_model_tokenizer(experiment_path, merge=False, device="cuda:0"):
-    cfg = load_config_yaml(os.path.join(experiment_path, "cfg.yaml"))
-    cfg.architecture.pretrained = False
-    cfg.architecture.gradient_checkpointing = False
-    cfg.environment._device = device.replace("_shard", "")
-    cfg.environment._local_rank = 0
-    cfg.prediction._visibility["num_history"] = 1
-
-    tokenizer = get_tokenizer(cfg)
-
-    gc.collect()
-    torch.cuda.empty_cache()
-
-    if (
-        merge
-        and cfg.training.lora
-        and cfg.architecture.backbone_dtype in ("int4", "int8")
-    ):
-        logger.info("Loading backbone in float16 for merging LORA weights.")
-        cfg.architecture.backbone_dtype = "float16"
-        cfg.architecture.pretrained = True
-
-    with torch.device(cfg.environment._device):
-        model = cfg.architecture.model_class(cfg)
-        cfg.architecture.pretrained_weights = os.path.join(
-            experiment_path, "checkpoint.pth"
-        )
-        load_checkpoint(cfg, model, strict=False)
-
-    if merge and cfg.training.lora:
-        # merges the LoRa layers into the base model.
-        # This is needed if one wants to use the base model as a standalone model.
-        logger.info("Merging LORA layers with base model.")
-        model.backbone = model.backbone.merge_and_unload()
-
-    if device == "cpu_shard":
-        max_memory = get_balanced_memory(
-            model,
-        )
-        device_map = infer_auto_device_map(model, max_memory=max_memory)
-        model = dispatch_model(
-            model,
-            device_map=device_map,
-        )
-
-    model = model.eval()
-    model.backbone.use_cache = True
-
-    return cfg, model, tokenizer
