+++ app_utils/utils.py
@@ -28,7 +28,6 @@ from datasets import load_dataset
 from h2o_wave import Q, ui
 from pandas.core.frame import DataFrame
 from sqlitedict import SqliteDict
-from transformers import AutoTokenizer, TextStreamer
 
 from app_utils.db import Experiment
 from llm_studio.python_configs.text_causal_language_modeling_config import (
@@ -1948,27 +1947,3 @@ def prepare_default_dataset(path):
     )
 
     return df_assistant[(df_assistant["rank"] == 0.0) & (df_assistant["lang"] == "en")]
-
-
-class WaveChatStreamer(TextStreamer):
-    def __init__(
-        self, tokenizer: AutoTokenizer, q: Q, text_cleaner=None, **decode_kwargs
-    ):
-        super().__init__(tokenizer, skip_prompt=True, **decode_kwargs)
-        self.text_cleaner = text_cleaner
-        self.answer = ""
-        self.loop = asyncio.get_event_loop()
-        self.q = q
-
-    def on_finalized_text(self, text: str, stream_end: bool = False):
-        self.answer += f" {text}"
-        if self.answer.endswith(self.tokenizer.eos_token):
-            # text generation is stopped
-            self.answer = self.answer.replace(self.tokenizer.eos_token, "")
-        if self.text_cleaner:
-            self.answer = self.text_cleaner(self.answer)
-        self.loop.create_task(self.push_to_chat())
-
-    async def push_to_chat(self):
-        self.q.page["experiment/display/chat"].data[-1] = [self.answer, False]
-        await self.q.page.save()
